pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.2",
    "pandas-ta>=0.3.14b0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "ta>=0.11.0",
    "yfinance>=0.2.65",
    "requests>=2.32.5",
    "ta-lib>=0.6.5",
]
